Изменены комплексные числа

+++ include/complex.h
@@ -4,11 +4,11 @@
 #include <stdio.h>
 
 typedef struct {
-    double real; //действительная часть
-    double imag; //мнимая часть
+    int real; //действительная часть
+    int imag; //мнимая часть
 } Complex;
 
-Complex complex_create(double real, double imag);
+Complex complex_create(int real, int imag);
 Complex complex_add(Complex a, Complex b);
 Complex complex_mul(Complex a, Complex b);
 void complex_print(Complex c);
